FLIZ-452 feat: /notification/connect 페이지 네트워크 최적화
